+++ apm-sniffer/apm-sdk-plugin/dbcp-2.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/dbcp/v2/PoolingSetUrlInterceptor.java
@@ -17,7 +17,7 @@
 
 package org.apache.skywalking.apm.plugin.dbcp.v2;
 
-import org.apache.commons.dbcp2.BasicDataSource;
+import org.apache.commons.dbcp2.BasicDataSourceMXBean;
 import org.apache.skywalking.apm.agent.core.logging.api.ILog;
 import org.apache.skywalking.apm.agent.core.logging.api.LogManager;
 import org.apache.skywalking.apm.agent.core.meter.MeterFactory;
@@ -34,21 +34,21 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
- * {@link PoolingSetUrlSourceInterceptor} intercepted the method of DBCP set url and register metric monitor.
+ * {@link PoolingSetUrlInterceptor} intercepted the method of DBCP set url and register metric monitor.
  */
-public class PoolingSetUrlSourceInterceptor implements InstanceMethodsAroundInterceptor {
+public class PoolingSetUrlInterceptor implements InstanceMethodsAroundInterceptor {
     private static final String METER_NAME = "datasource";
-    private static final ILog LOGGER = LogManager.getLogger(PoolingSetUrlSourceInterceptor.class);
-    private static final Map<String, Function<BasicDataSource, Supplier<Double>>> METRIC_MAP = new HashMap<String, Function<BasicDataSource, Supplier<Double>>>();
+    private static final ILog LOGGER = LogManager.getLogger(PoolingSetUrlInterceptor.class);
+    private static final Map<String, Function<BasicDataSourceMXBean, Supplier<Double>>> METRIC_MAP = new HashMap<String, Function<BasicDataSourceMXBean, Supplier<Double>>>();
 
     static {
-        METRIC_MAP.put("numActive", (BasicDataSource basicDataSource) -> () -> (double) basicDataSource.getNumActive());
-        METRIC_MAP.put("maxTotal", (BasicDataSource basicDataSource) -> () -> (double) basicDataSource.getMaxTotal());
-        METRIC_MAP.put("numIdle", (BasicDataSource basicDataSource) -> () -> (double) (basicDataSource.getNumIdle()));
-        METRIC_MAP.put("maxWaitMillis", (BasicDataSource basicDataSource) -> () -> (double) basicDataSource.getMaxWaitMillis());
-        METRIC_MAP.put("maxIdle", (BasicDataSource basicDataSource) -> () -> (double) basicDataSource.getMaxIdle());
-        METRIC_MAP.put("minIdle", (BasicDataSource basicDataSource) -> () -> (double) basicDataSource.getMinIdle());
-        METRIC_MAP.put("initialSize", (BasicDataSource basicDataSource) -> () -> (double) basicDataSource.getInitialSize());
+        METRIC_MAP.put("numActive", (BasicDataSourceMXBean basicDataSource) -> () -> (double) basicDataSource.getNumActive());
+        METRIC_MAP.put("maxTotal", (BasicDataSourceMXBean basicDataSource) -> () -> (double) basicDataSource.getMaxTotal());
+        METRIC_MAP.put("numIdle", (BasicDataSourceMXBean basicDataSource) -> () -> (double) (basicDataSource.getNumIdle()));
+        METRIC_MAP.put("maxWaitMillis", (BasicDataSourceMXBean basicDataSource) -> () -> (double) basicDataSource.getMaxWaitMillis());
+        METRIC_MAP.put("maxIdle", (BasicDataSourceMXBean basicDataSource) -> () -> (double) basicDataSource.getMaxIdle());
+        METRIC_MAP.put("minIdle", (BasicDataSourceMXBean basicDataSource) -> () -> (double) basicDataSource.getMinIdle());
+        METRIC_MAP.put("initialSize", (BasicDataSourceMXBean basicDataSource) -> () -> (double) basicDataSource.getInitialSize());
     }
 
     @Override
@@ -61,7 +61,7 @@ public class PoolingSetUrlSourceInterceptor implements InstanceMethodsAroundInte
         if (LOGGER.isInfoEnable()) {
             LOGGER.info("metric dbcp init");
         }
-        BasicDataSource basicDataSource = (BasicDataSource) objInst;
+        BasicDataSourceMXBean basicDataSource = (BasicDataSourceMXBean) objInst;
         ConnectionInfo connectionInfo = URLParser.parser((String) allArguments[0]);
         String tagValue = connectionInfo.getDatabaseName() + "_" + connectionInfo.getDatabasePeer();
         METRIC_MAP.forEach((key, value) -> MeterFactory.gauge(METER_NAME, value.apply(basicDataSource))